requirements.txt
gym==0.10.9
numpy==1.19.5
tensorboard==1.14.0
tensorboard-logger==0.1.0
tensorflow==1.14.0
PyYAML==5.4.1


